pyproject.toml
[project]
name = "mcplibrarian"
version = "0.1.0"
description = "Automatically wrap MCP servers in Docker containers with lazy loading"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "toml>=0.10.2",
    "docker>=7.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcplibrarian = "mcp_dockerize.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_dockerize"]
